backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# AI/ML
anthropic==0.18.1

# Environment & Config
python-dotenv==1.0.1

# AWS Integration
boto3==1.34.34
django-storages==1.14.2

# Market Data APIs
requests==2.31.0

# PDF Processing (for Phase 1 priority)
PyPDF2==3.0.1
pdfplumber==0.10.4
pypdf==4.0.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
ipython==8.21.0
